+++ src/main/java/org/recap/service/email/datadump/DataDumpEmailService.java
@@ -25,7 +25,7 @@ public class DataDumpEmailService {
     @Autowired
     private ProducerTemplate producer;
 
-    public void sendEmail(List<String> institutionCodes, Long totalRecordCount, String requestingInstitutionCode, String transmissionType, String dateTimeStringForFolder, String toEmailAddress) {
+    public void sendEmail(List<String> institutionCodes, Integer totalRecordCount, String requestingInstitutionCode, String transmissionType, String dateTimeStringForFolder, String toEmailAddress) {
         EmailPayLoad emailPayLoad = new EmailPayLoad();
         emailPayLoad.setInstitutions(institutionCodes);
         emailPayLoad.setLocation(getLocation(transmissionType, requestingInstitutionCode,dateTimeStringForFolder));
